requirements.txt
asgiref==3.10.0
Django==5.2.7
django-seed==0.3.1
djangorestframework==3.16.1
factory_boy==3.3.3
Faker==37.11.0
pillow==11.3.0
sqlparse==0.5.3
toposort==1.10
tzdata==2025.2
